+++ nemdata/cli.py
@@ -18,10 +18,15 @@ from nemdata.nemde import download_nemde
 )
 @click.option("--start", "-s", default="2018-01", help="start date (YYYY-MM)")
 @click.option("--end", "-e", default="2018-03", help="end date (incusive) (YYYY-MM)")
-def cli(start: str, end: str, report: str) -> None:
+@click.option(
+    "--dry-run/--no-dry-run",
+    default=False,
+    help="whether to save downloaded data to disk",
+)
+def cli(start: str, end: str, table: str, dry_run: bool) -> None:
     """nemdata is a tool to access NEM data from AEMO."""
     print(":wave: from nemdata\n")
-    download(start, end, report)
+    download(start, end, table, dry_run=dry_run)
 
 
 def download(
@@ -29,6 +34,7 @@ def download(
     end: str,
     table_name: str,
     base_directory: pathlib.Path = DEFAULT_BASE_DIR,
+    dry_run: bool = False,
 ) -> pd.DataFrame:
     print(f"[bold green]Downloader[/]: table: {table_name}")
     tables: dict[str, typing.Callable] = {
@@ -39,7 +45,11 @@ def download(
         "interconnectors": mmsdm.download_mmsdm,
     }
     return tables[table_name](
-        start, end, table_name=table_name, base_directory=base_directory
+        start,
+        end,
+        table_name=table_name,
+        base_directory=base_directory,
+        dry_run=dry_run,
     )