+++ nemdata/nemde.py
@@ -81,6 +81,7 @@ def download_nemde(
     end: str,
     table_name: str = "nemde",
     base_directory: pathlib.Path = DEFAULT_BASE_DIRECTORY,
+    dry_run: bool = False,
 ) -> pd.DataFrame:
     """main for downloading MMSDMFiles"""
     table = NEMDETable()
@@ -103,9 +104,10 @@ def download_nemde(
             data["PeriodID"] = pd.to_datetime(data["PeriodID"]).dt.tz_localize(None)
             data = utils.add_interval_column(data, table)
 
-            print(f" [green]SAVING [/] {clean_fi}")
-            data.to_csv(clean_fi.with_suffix(".csv"))
-            data.to_parquet(clean_fi.with_suffix(".parquet"))
+            if not dry_run:
+                print(f" [green]SAVING [/] {clean_fi}")
+                data.to_csv(clean_fi.with_suffix(".csv"))
+                data.to_parquet(clean_fi.with_suffix(".parquet"))
         dataset.append(data)
     return pd.concat(dataset, axis=0)