+++ changelog.md
@@ -2,11 +2,13 @@
 
 0.2.0
 - breaking change moving from concept of `report` to concept of `table`,
-- CLI option change from `-r --report` to `-t` `--table`,
+- breaking CLI option change from `-r --report` to `-t` `--table`,
+- added `--dry-run` to CLI,
+- added changelog,
 - add `pydantic`, `black` and `isort`,
 - introduce `MMSDMTable` and `MMSDMFile`,
 - add `make check`, `make lint` and `make static`,
-- added static typing checks,
+- added static typing check to CI,
 - added `frequency_minutes` as a column in `clean.parquet`.
 
 0.1.4 - cleanup of unused code.